requirements.txt
pandas
numpy
tqdm
opencv-python
matplotlib
numba
wand
Pillow
requests
pdf2image
gliner
seaborn
scikit-learn
scipy
textstat
datasets
mistralai
python-dotenv
jiwer
pyarrow
transformers
torch
pytesseract
huggingface-hub
doclayout-yolo
litellm
tenacity